Removed comments and minor fix to style
